Define all basic functions in a different header
